requirements/requirements.txt
Django>=4.2
djangorestframework>=3.14

paddleocr>=2.7
paddlepaddle>=2.5.0

pdf2image>=1.16.3
pdfplumber>=0.10.0

Pillow>=10.0.0
opencv-python>=4.8.0

requests>=2.31.0
python-dotenv>=1.0.0

numpy>=1.24.0

# OCR — Tesseract (fast, recommended) or PaddleOCR (fallback)
pytesseract>=0.3.10
